+++ templates/pageComponents/componentName.tsx
@@ -1,4 +1,5 @@
 import { Component } from "@/components/styled/content"
+
 import type { TCOMPONENTNAME } from "./types"
 
 export default function COMPONENTNAME(props: TCOMPONENTNAME) {